backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
pydantic[email]==2.5.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
openai==1.3.5
httpx==0.25.2
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
PyJWT==2.8.0
bcrypt==4.1.2
supabase==2.3.4
psycopg2-binary==2.9.9
cryptography==41.0.7
requests==2.31.0